feat: provide property for client components to use server action
