+++ src/app/[locale]/auth/reset/page.tsx
@@ -1,5 +1,6 @@
 import { useTranslations } from "next-intl";
 import { ResetForm } from "./form";
+import { resetPassword } from "@/lib/actions";
 
 export default function Page() {
   const t = useTranslations("auth.reset_form");
@@ -7,7 +8,7 @@ export default function Page() {
   return (
     <>
       <h1>{t("heading")}</h1>
-      <ResetForm />
+      <ResetForm resetPassword={resetPassword} />
     </>
   );
 }
