+++ src/app/[locale]/auth/signin/page.tsx
@@ -1,5 +1,6 @@
 import { useTranslations } from "next-intl";
 import { SigninForm } from "./form";
+import { signIn } from "@/lib/actions";
 
 export default function Page() {
   const t = useTranslations("auth.signin_form");
@@ -7,7 +8,7 @@ export default function Page() {
   return (
     <>
       <h1>{t("heading")}</h1>
-      <SigninForm />
+      <SigninForm signIn={signIn} />
     </>
   );
 }
