+++ src/app/[locale]/auth/signup/page.tsx
@@ -1,5 +1,6 @@
 import { useTranslations } from "next-intl";
 import { SignupForm } from "./form";
+import { signUp } from "@/lib/actions";
 
 export default function Page() {
   const t = useTranslations("auth.signup_form");
@@ -7,7 +8,7 @@ export default function Page() {
   return (
     <>
       <h1>{t("heading")}</h1>
-      <SignupForm />
+      <SignupForm signUp={signUp} />
     </>
   );
 }
