--- src/lib/actions/auth/login.ts
@@ -1,6 +1,0 @@
-import z from "zod";
-
-export const loginSchema = z.object({
-  email: z.string().email(),
-  password: z.string().min(8),
-});
