+++ src/lib/validations/auth/reset.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+import { zfd } from "zod-form-data";
+
+export const resetFormSchema = zfd.formData({
+  email: zfd.text(z.string().email()),
+});
+
+export type ResetFormInput = z.infer<typeof resetFormSchema>;
